// src/app/modules/usuario.interface.ts
export interface Usuario {
    _id : string;
    correo: string;
    pais: string;
    departamento: string;
    ciudad: string;
    telefono1: string;
    telefono2: string;
    nombreEntrega: string;
    direccion: string;
    codigoPostal: string;
    detalle: string;
    instruccionesEntrega: string;
    createdAt: Date;
}